Working test scene!!
Organized the videos into a working test scene/the start of the game. Buttons are still invisible however they are working and in the middle of the screen. Loading screen is annoying but I don't want to risk breaking everything trying to import videos directly into the project, as I believe that is where my first attempt went wrong.